feat: product delete 조건 추가
다른 제품의 sub assay로 등록되어 있으면 삭제 불가

+++ src/main/java/com/pororoz/istock/domain/product/service/ProductService.java
@@ -57,6 +57,9 @@ public class ProductService {
   public ProductServiceResponse deleteProduct(Long productId) {
     Product product = productRepository.findById(productId)
         .orElseThrow(ProductNotFoundException::new);
+    if (bomRepository.existsByProductNumber(product.getProductNumber())) {
+      throw new RegisteredAsSubAssayException();
+    }
     productRepository.delete(product);
     return ProductServiceResponse.of(product);
   }
